refactor: More modular file structure
