+++ client/src/components/HabitForm/Buttons/Buttons.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Button, ButtonGroup } from '../Shared/Forms/Form.styled';
+import { Button, ButtonGroup } from '../../Shared/Forms/Form.styled';
 
 export const LeftButtons = ({ step, setStep }) => {
 	if (step > 0) return <Button onClick={() => setStep(step - 1)}>Back</Button>;
