+++ client/src/components/HabitForm/Content/Content.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { ContentArea, Row } from './HabitForm.styled';
-import FormGroup from '../Shared/Forms/FormGroup';
+import { ContentArea, Row } from '../HabitForm.styled';
+import FormGroup from '../../Shared/Forms/FormGroup';
 
 const Content = ({
 	step,
